requirements.txt
streamlit
pandas
azure-eventhub
python-dotenv
requests
streamlit-autorefresh 